backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
pydantic
python-multipart
httpx
pyyaml
pytest
pytest-asyncio
